reestructured and continue with separation of logic in order that it can be used as a library
