+++ error/error.go
@@ -1,0 +1,38 @@
+package error
+
+import (
+	"encoding/json"
+	"fmt"
+	"github.com/Sirupsen/logrus"
+	"net/http"
+)
+
+var log = logrus.New()
+
+// APIErrorMessage is an object that defines when a generic error occurred
+type APIErrorMessage struct {
+	Status string
+	Error  string
+}
+
+// HandleError is a generic error handler
+func HandleError(tag string, errorMsg string, rawErr error, code int, w http.ResponseWriter, r *http.Request) {
+	log.WithFields(logrus.Fields{
+		"prefix":   tag,
+		"errorMsg": errorMsg,
+	}).Error(rawErr)
+
+	errorObj := APIErrorMessage{"error", errorMsg}
+	responseMsg, err := json.Marshal(&errorObj)
+
+	if err != nil {
+		log.WithField("prefix", tag).Error("[Error Handler] Couldn't marshal error stats: ", err)
+		fmt.Fprintf(w, "System Error")
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	fmt.Fprintf(w, string(responseMsg))
+}
+
